Backend/requirements.txt
# FastAPI framework and server
fastapi
uvicorn[standard]
python-jose[cryptography]

# Database ORM and adapter
sqlalchemy
psycopg2-binary
alembic

# Environment variable management and Pydantic settings
pydantic-settings
python-dotenv

# Password hashing (will be used in auth.py)
passlib 
bcrypt==4.0.1

pydantic[email]

#AI
google-generativeai

#Frontend
chart.js react-chartjs-2